29 add from state (#54)
* Added from_state

Created a from_state function which takes a list of qubits and creates a state.

* Added error handling and comments

from_state now panics when an input is incorrect. Will switch to recoverable errors.

* Implemented recoverable errors

Instead of using assert to handle errors Result is used to handle them.

* Fix bug

* Added test

* Added from_state

Created a from_state function which takes a list of qubits and creates a state.

* Added error handling and comments

from_state now panics when an input is incorrect. Will switch to recoverable errors.

* Implemented recoverable errors

Instead of using assert to handle errors Result is used to handle them.

* Fix bug

* Added test

* Fix out of scope error

* Remove unnecessary mut

* Fix request and update from stable

Fixed requested changes and merged with updated from stable branch